The reflection tool will automatically be available (part of ReAct's design is to have such a component) through shared tools and it being default added to the list of tools within the BaseReAct agent template
